1111111web-service/requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
redis==5.0.1 